requirements.txt
mujoco-py 
pygame 
networkx 
dotmap
git+https://github.com/Farama-Foundation/d4rl@master#egg=d4rl
matplotlib
gym
ray
tensorboard
pytorch-mppi[tune]
arm-pytorch-utilities
pandas==2.0.0
numpy==1.24.3
neupy
